requirements.txt
nicegui
requests
boto3
rasterio
numpy
shapely
prefect
pyproj
pillow
matplotlib